+++ spec/spec_helper.rb
@@ -14,12 +14,11 @@
     limitations under the License.
 =end
 
+# Uncomment to show output from the Framework.
+require_relative '../lib/arachni/ui/cli/output'
 require_relative '../lib/arachni'
 require_relative '../lib/arachni/processes'
 require_relative '../lib/arachni/processes/helpers'
-
-# Uncomment to show output from the Framework.
-#require_relative '../lib/arachni/ui/cli/output'
 
 require_relative 'support/helpers/paths'
 require_relative 'support/helpers/requires'